foreign keys should be bigint

+++ db/migrate/20210727063853_create_fleet_travel_logs.rb
@@ -2,8 +2,8 @@ class CreateFleetTravelLogs < ActiveRecord::Migration[6.0]
   def change
     create_table :fleet_travel_logs do |t|
       t.references :ship, index: true, foreign_key: true
-      t.integer :from_port, foreign_key: true
-      t.integer :to_port, foreign_key: true
+      t.bigint :from_port, foreign_key: true
+      t.bigint :to_port, foreign_key: true
       t.datetime :time_depart
       t.datetime :time_arrive
